Refactor Dockerfile to improve readability and streamline installation process

+++ Dockerfile
@@ -1,21 +1,24 @@
 FROM python:3.12-slim-bookworm
 # Update package lists and install required tools
-RUN apt-get -y update && \
-    apt-get -y install --no-install-recommends \
-    curl \
-    gnupg \
-    apt-transport-https \
-    net-tools \
-    tini && \
-    apt-get clean && rm -rf /var/lib/apt/lists/*
+RUN <<EOF
+apt-get -y update && \
+apt-get -y install --no-install-recommends \
+curl \
+gnupg \
+apt-transport-https \
+net-tools \
+tini && \
+
 # Add Microsoft package repository and install ODBC 18
 # https://learn.microsoft.com/en-us/sql/connect/odbc/linux-mac/installing-the-microsoft-odbc-driver-for-sql-server?view=sql-server-ver16&tabs=debian18-install%2Calpine17-install%2Cdebian8-install%2Credhat7-13-install%2Crhel7-offline
-#RUN OS_VERSION=$(grep VERSION_ID /etc/os-release | cut -d '"' -f 2 | cut -d '.' -f 1) && \
-#    curl -sSL -O https://packages.microsoft.com/config/debian/${OS_VERSION}/packages-microsoft-prod.deb && \
-#    dpkg -i packages-microsoft-prod.deb && rm packages-microsoft-prod.deb && \
-#    apt-get update -y && \
-#    ACCEPT_EULA=Y apt-get install --no-install-recommends -y msodbcsql18 procps net-tools tini && \
-#    apt-get clean && rm -rf /var/lib/apt/lists/*
+OS_VERSION=$(grep VERSION_ID /etc/os-release | cut -d '"' -f 2 | cut -d '.' -f 1) && \
+#curl -sSL -O https://packages.microsoft.com/config/debian/${OS_VERSION}/packages-microsoft-prod.deb && \
+#dpkg -i packages-microsoft-prod.deb && rm packages-microsoft-prod.deb && \
+#apt-get update -y && \
+#ACCEPT_EULA=Y apt-get install --no-install-recommends -y msodbcsql18 procps net-tools tini
+
+apt-get clean && rm -rf /var/lib/apt/lists/*
+EOF
 
 
 # https://docs.python.org/3/using/cmdline.html#envvar-PYTHONDONTWRITEBYTECODE
@@ -30,7 +33,7 @@ ENV PIP_DEFAULT_TIMEOUT=100
 
 # Install uv
 ADD https://astral.sh/uv/install.sh /install.sh
-RUN chmod -R 755 /install.sh && /install.sh && rm /install.sh
+RUN chmod 755 /install.sh && /install.sh && rm /install.sh
 
 ENV PATH=/root/.local/bin:$PATH
 
@@ -44,8 +47,9 @@ RUN uv pip install -r pyproject.toml --system
 USER appuser
 # Now copy the app
 COPY --chown=appuser:appuser ./app .
-
-RUN cat <<____HERE > .streamlit/secrets.toml
+RUN <<EOF
+python -c "import uuid; print(uuid.uuid4())" > auth_cookie_key.uuid
+cat <<____HERE > .streamlit/secrets.toml
 
 [ldap]
 # See https://www.forumsys.com/2022/05/10/online-ldap-test-server/
@@ -65,7 +69,7 @@ remember_me = "login_remember_me"
 
 [auth_cookie]
 name = "stusermanagerdemo"
-key = "hf9u9iXyZrUAUn4jue2!oDz_96dtAX*Q"
+key = "$(cat auth_cookie_key.uuid)"
 expiry_days = 1
 auto_renewal = true
 delay_sec = 0.1
@@ -76,13 +80,16 @@ keyName = "authkey"
 
 ____HERE
 
-RUN chmod +x run.sh \
-    && rm -f uv.lock \
-    && rm -rf logs \
-    && mkdir logs \
-    && mkdir -p .rsa \
-    && python3 generateKeys.py && chmod 0600 .rsa/authkey
-
+# if the grep fails, the file was not created correct. Run the job again
+echo "If the following grep fails, the secrets.toml file was not created. Build the image again"
+grep keyName .streamlit/secrets.toml || exit 1
+chmod +x run.sh \
+&& rm -f uv.lock \
+&& rm -rf logs \
+&& mkdir logs \
+&& mkdir -p .rsa \
+&& python3 generateKeys.py && chmod 0600 .rsa/authkey
+EOF
 ENV PYTHONPATH=/app
 ENV PYTHONFAULTHANDLER=1
